`tools/data-api-repository`: moving the repository implementation into the repository module
This allows other things to use it, with tweaks, as needed.
