+++ tools/data-api-repository/repository/remove_service.go
