+++ tools/data-api-repository/repository/stages/service_definition.go
@@ -5,12 +5,12 @@ package stages
 
 import (
 	"fmt"
-	"github.com/hashicorp/go-hclog"
 	"path/filepath"
 
+	"github.com/hashicorp/go-hclog"
+	"github.com/hashicorp/pandora/tools/data-api-repository/repository/helpers"
+	"github.com/hashicorp/pandora/tools/data-api-repository/repository/transforms"
 	"github.com/hashicorp/pandora/tools/data-api-sdk/v1/models"
-	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/repository/helpers"
-	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/repository/transforms"
 )
 
 var _ Stage = ServiceDefinitionStage{}
