+++ tools/data-api-repository/repository/stages/terraform_resource_tests.go
@@ -8,8 +8,8 @@ import (
 	"path/filepath"
 
 	"github.com/hashicorp/go-hclog"
+	"github.com/hashicorp/pandora/tools/data-api-repository/repository/helpers"
 	"github.com/hashicorp/pandora/tools/data-api-sdk/v1/models"
-	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/repository/helpers"
 )
 
 var _ Stage = TerraformResourceTestsStage{}
