+++ tools/data-api-repository/repository/transforms/terraform_resource_definition.go
