+++ tools/data-api-repository/repository/transforms/terraform_schema_object_definition.go
